chore: renamed vars and interfaces
