+++ packages/scaffolding-cli/templates/src/csr/src/containers/home.tsx
@@ -1,15 +1,15 @@
 import React, {useCallback} from "react"
 import {useDispatch, useSelector} from "react-redux"
 import RestaurantListComponent from "../components/RestaurantList"
-import {IApplicationState} from "../state/ducks/index"
-import {fetchPosts} from "../state/ducks/menus/actions"
-import {IPostState} from "../state/ducks/menus/types"
+import {ApplicationState} from "../state/ducks/index"
+import {fetchPosts} from "../state/ducks/get-menus/actions"
+import {IPostState} from "../state/ducks/get-menus/types"
 
 const PostListContainer = () => {
     const dispatch = useDispatch()
 
     const stateToProps: IPostState = useSelector(
-        ({post}: IApplicationState) => ({
+        ({post}: ApplicationState) => ({
             loading: post.loading,
             errors: post.errors,
             data: post.data,
