+++ packages/scaffolding-cli/templates/src/csr/src/state/ducks/get-menus/actions.ts
@@ -1,13 +1,13 @@
 import {action} from "typesafe-actions"
 /* eslint import/no-cycle: [2, { maxDepth: 1 }] */
-import {IPostRaw, PostActionTypes} from "./types"
+import {IPostRaw, GetMenuActionTypes} from "./types"
 
 export const fetchPosts = () =>
-    action(PostActionTypes.FETCH_POSTS, [], {
+    action(GetMenuActionTypes.FETCH_POSTS, [], {
         method: "get",
         route: "/menu",
     })
 export const fetchPostsSuccess = (data: IPostRaw[]) =>
-    action(PostActionTypes.FETCH_POSTS_SUCCESS, data)
+    action(GetMenuActionTypes.FETCH_POSTS_SUCCESS, data)
 export const fetchPostsError = (message: string) =>
-    action(PostActionTypes.FETCH_POSTS_ERROR, message)
+    action(GetMenuActionTypes.FETCH_POSTS_ERROR, message)
